Move debian builder functions
